+++ lizmap/server_lwc.py
@@ -628,7 +628,6 @@ class ServerManager:
         # Markdown
         markdown = '**Versions :**\n\n'
         markdown += '* Lizmap Web Client : {}\n'.format(lizmap_version)
-        markdown += '* Lizmap Web Client target version : {}\n'.format(self.parent.current_lwc_version().value)
         markdown += '* Lizmap plugin : {}\n'.format(version())
         markdown += '* QGIS Desktop : {}\n'.format(Qgis.QGIS_VERSION.split('-')[0])
         qgis_cell.setData(Qt.UserRole, markdown)