+++ CHANGELOG.md
@@ -5,9 +5,10 @@
 * Check for space in the Lizmap URL before requesting metadata
 * Remove the target version dropdown menu, it will now use the server dropdown menu
 * The server must have a valid status before using the plugin
-* The metadata for each server is stored for maximum two days
+* The metadata for each server is stored for maximum seven days
 * Fix the shortname generation when the layer name contains a `-`
 * Fix the last used repository used in the dropdown menu
+* Fix the new thumbnail checker
 
 ## 3.11.5 - 2023-04-11